Implemented a value function approximator
Value function approximators play a role as the action value function,
which returns the q(state, action) with the given state(in this case
observationSequence) and the given action.
The implemented approximator care only the last observation
in order to meet with the concept of P-controller(ConcAgent).
This specification should be expanded in this routine.