Added the regularization term on the reward
Without regularization term, such as abs(u), the training process is
likely to be unstable. For example, the magnitude of u might become
a large negative value and DO converges 0. It can cause the difficulty
of the training. The regularization might be beneficial because it's
expected that u remains around 0.