Revised the subsection3.4(casestudy#4a)
In the case study #4a,
the cases with fixPolicySd = False or True had been compared
and its consequence had said that the case of fixed sd: "fixPolicySd = 1"
was better, in the previous revision.

This was error, in fact,
the case of tunable sd: "fixPolicySd = 0" was better.

The explanation was revised so that it could meet with the result.