Added the note of case study 2a
Case study 2 will investigate the RL parameter to control the
responsiveness of the controller.

Compare the following case in the case study 2a:
a). gamma = 0.5, with the short time horizon of reward
b). gamma = 0.9, with the medium time horizon of reward
c). gamma = 0.975, with the long time horizon of reward