Added the notes for case study 3c
The case study 3c studies how the parameter: weight on error
(regularization term) affects the learning and the performance of agents.
More precisely say that we study the following effects of weightOnError
* the effect on the responsiveness against the stepwise disturbance,
* and the effect on the learning of the gaussian distributed disturbance.

Compare the cases as follow:
1). weightOnError = 0.99, learning with small regularization,
2). weightOnError = 0.90, learning with fairly small regularization,
3). weightOnError = 0.50, learning under the situation where the
  regularization term eqauls with the error.

In all the cases, agents are learned under the environments with and without
small gaussian disturbance, respectively, which means that 6 cases in total
are compared.